pyproject.toml
[project]
name = "nonebot-plugin-bittorrents"
version = "1.2.1"
description = "适配NoneBot的磁力搜索插件，通过机器人帮你寻找电影、软件或者学习资料"
authors = [
    {name = "NightDust981989"}
]
license = "MIT"
readme = "README.md"
dependencies = [
    "httpx>=0.23.0",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
    "nonebot2>=2.3.0",
]

[project.urls]
homepage = "https://github.com/NightDust981989/nonebot-plugin-BitTorrents"
repository = "https://github.com/NightDust981989/nonebot-plugin-BitTorrents"

[build-system]
requires = ["setuptools>=61.0","wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["nonebot_plugin_BitTorrents*"]

[tool.nonebot]
plugins = ["nonebot_plugin_BitTorrents"] 